[refactor](be) Remove redundant casts from BE code
### What problem does this PR solve?

Issue Number: None

Related PR: None

Problem Summary: Remove redundant casts reported by clang-tidy readability-redundant-casting in BE code. These casts were no-op conversions to the same type, including redundant static_cast, reinterpret_cast, and C-style casts. The change keeps casts that still perform real type conversion and only removes the redundant outer casts reported by clang-tidy.

### Release note

None

### Check List (For Author)

- Test: Static analysis / Build
    - Ran clang-tidy readability-redundant-casting on the modified BE files; no remaining readability-redundant-casting warnings were reported.
    - Ran clang-format-16 on modified BE files.
    - Ran git diff --check -- be/src.
    - Ran sh build.sh --be -j48, build succeeded with "Successfully build Doris".
- Behavior changed: No
- Does this need documentation: No

+++ be/src/format/orc/vorc_reader.cpp
@@ -747,7 +747,7 @@ std::tuple<bool, orc::Literal> convert_to_orc_literal(const orc::Type* type,
             } else if constexpr (primitive_type == TYPE_INT) {
                 return std::make_tuple(true, orc::Literal(int64_t(*((int32_t*)value))));
             } else if constexpr (primitive_type == TYPE_BIGINT) {
-                return std::make_tuple(true, orc::Literal(int64_t(*((int64_t*)value))));
+                return std::make_tuple(true, orc::Literal(*((int64_t*)value)));
             }
             return std::make_tuple(false, orc::Literal(false));
         }
@@ -755,7 +755,7 @@ std::tuple<bool, orc::Literal> convert_to_orc_literal(const orc::Type* type,
             if constexpr (primitive_type == TYPE_FLOAT) {
                 return std::make_tuple(true, orc::Literal(double(*((float*)value))));
             } else if constexpr (primitive_type == TYPE_DOUBLE) {
-                return std::make_tuple(true, orc::Literal(double(*((double*)value))));
+                return std::make_tuple(true, orc::Literal(*((double*)value)));
             }
             return std::make_tuple(false, orc::Literal(false));
         }
